Refactor ACES provisioner for Sonar
